Small re-org of the code
